提交字体配置表工具完善

+++ Framework_UClient/Assets/_FutureFrame/FutureCore/Data/Veriy/HotFixVerify.cs
@@ -12,7 +12,7 @@ namespace FutureCore.Data
         /// <summary>
         /// 版本号
         /// </summary>
-        public int version;
+        public ulong version;
         /// <summary>
         /// 最新的打包时间
         /// </summary>
